abstracted document into parent class
now text document and image documents are derived from ArchiveDocument
