+++ Assets/Scripts/Archive System/ImageDocument.cs
@@ -1,0 +1,10 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+[CreateAssetMenu(fileName = "Archive Image", menuName = "ScriptableObjects/ArchiveImage")]
+public class ImageDocument : ArchiveDocument
+{
+    public Image documentImage;
+}
