+++ Assets/Scripts/Archive System/ImageDocument.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 37ddb04a8306dfa48b0d0e0ba94a3bba
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
