+++ Assets/Scripts/Archive System/TextDocument.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b5e53cddb7866b341b0b794ba919c5fd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
